requirements.txt
Flask==2.1.3
python-dotenv==1.0.0
openai==0.27.6
PyPDF2==3.0.0
werkzeug==2.0.3
